requirements.txt
flask==3.1.1
flask-cors==4.0.0
openai==1.54.0
google-generativeai==0.8.3
speechrecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
python-multipart==0.0.6
gunicorn==21.2.0
# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.25
chromadb==0.4.22
sentence-transformers==2.5.1
faiss-cpu==1.7.4
tiktoken==0.6.0 